LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-top-layer-nesting-anchor.tentative-expected.txt

Single popover=auto ancestor
Single popover=manual ancestor
Nested popover=auto ancestors
Nested popover=auto ancestors, target is outer
Top layer inside of nested element

FAIL Single popover=auto ancestor with dialog assert_equals: Incorrect behavior expected true but got false
PASS Single popover=auto ancestor with dialog, top layer element *is* a popover
FAIL Single popover=auto ancestor with dialog, anchor attribute assert_equals: Incorrect behavior expected true but got false
FAIL Single popover=auto ancestor with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Single popover=auto ancestor with fullscreen, top layer element *is* a popover
FAIL Single popover=auto ancestor with fullscreen, anchor attribute promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Single popover=manual ancestor with dialog assert_equals: Incorrect behavior expected true but got false
PASS Single popover=manual ancestor with dialog, top layer element *is* a popover
FAIL Single popover=manual ancestor with dialog, anchor attribute assert_equals: Incorrect behavior expected true but got false
FAIL Single popover=manual ancestor with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Single popover=manual ancestor with fullscreen, top layer element *is* a popover
FAIL Single popover=manual ancestor with fullscreen, anchor attribute promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Nested popover=auto ancestors with dialog assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors with dialog, top layer element *is* a popover
FAIL Nested popover=auto ancestors with dialog, anchor attribute assert_equals: Incorrect behavior expected true but got false
FAIL Nested popover=auto ancestors with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors with fullscreen, top layer element *is* a popover
FAIL Nested popover=auto ancestors with fullscreen, anchor attribute promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Nested popover=auto ancestors, target is outer with dialog assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors, target is outer with dialog, top layer element *is* a popover
FAIL Nested popover=auto ancestors, target is outer with dialog, anchor attribute assert_equals: Incorrect behavior expected true but got false
FAIL Nested popover=auto ancestors, target is outer with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Nested popover=auto ancestors, target is outer with fullscreen, top layer element *is* a popover
FAIL Nested popover=auto ancestors, target is outer with fullscreen, anchor attribute promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Top layer inside of nested element with dialog assert_equals: Incorrect behavior expected true but got false
PASS Top layer inside of nested element with dialog, top layer element *is* a popover
FAIL Top layer inside of nested element with dialog, anchor attribute assert_equals: Incorrect behavior expected true but got false
FAIL Top layer inside of nested element with fullscreen assert_equals: Incorrect behavior expected true but got false
PASS Top layer inside of nested element with fullscreen, top layer element *is* a popover
FAIL Top layer inside of nested element with fullscreen, anchor attribute promise_test: Unhandled rejection with value: object "TypeError: Type error"
